css added. store data printed to screen. Needs to be formatted properly.
