+++ shells/browser/shared/view/App.jsx
@@ -1,8 +1,5 @@
 /** @format */
 import React, { useEffect, useState } from "react";
-import RightContainer from './RightContainer';
-import StoreDisplayer from './StoreDisplayer';
-import ComponentsRenderer from './ComponentsRenderer';
 
 const port = chrome.runtime.connect({ name: "test" });
 
@@ -49,20 +46,7 @@ const App = () => {
   // const test = logger('hi marc')
 
   return (
-
-    <div id="main-container">
-      <div className="columns">
-        {/* <div className="column">
-          <ComponentsRender />
-        </ div> */}
-        <div className="column border">
-          <StoreDisplayer />
-        </div>
-        <div className="column">
-          <RightContainer />
-        </div>
-      </div>
-    </div>
+    <p>App.jsx is loaded</p>
   );  
 };
