+++ shells/browser/shared/view/styles.scss
@@ -28,3 +28,11 @@ body {
   color: #5E72E4;
   font-weight: bold;
 }
+
+.nestedObject {
+  border: 1px solid black;
+}
+
+.records {
+  margin-left: 1em;
+}
