+++ src/devtools/DevTools.js
@@ -16,6 +16,10 @@ import React, { useState, useCallback, useEffect } from 'react';
 import type { FrontendBridge } from 'src/bridge';
 import Store from './store';
 import { BridgeContext, StoreContext } from './context';
+
+import RightContainer from './view/RightContainer';
+import StoreDisplayer from './view/StoreDisplayer';
+// import ComponentsRenderer from './ComponentsRenderer';
 // import Network from './Network/Network';
 // import StoreInspector from './StoreInspector/StoreInspector';
 // import TabBar from './TabBar';
@@ -114,9 +118,20 @@ export default function DevTools({
     <BridgeContext.Provider value={bridge}>
       <StoreContext.Provider value={store}>
         <p>Loaded context</p>
-        {allRecords}
+        {/* {allRecords} */}
+        <div className="column border">
+          <StoreDisplayer />
+        </div>
       </StoreContext.Provider>
     </BridgeContext.Provider>
   );
 }
 
+        // <div id="main-container">
+        //   <div className="columns">
+
+        //     <div className="column">
+        //       <RightContainer />
+        //     </div>
+        //   </div>
+        // </div>;
