+++ src/devtools/view/Components/Navbar.js
