+++ src/devtools/view/Components/StoreTimeline.js
