+++ src/devtools/view/ComponentsRenderer.js
