+++ src/devtools/view/RightContainer.js
